Extract NoCaretTextField to TestUtils.kt
