+++ compose/ui/ui/src/desktopTest/kotlin/androidx/compose/ui/TextFieldUndoTest.kt
@@ -16,39 +16,19 @@
 
 package androidx.compose.ui
 
-import androidx.compose.foundation.text.BasicTextField
-import androidx.compose.material.LocalTextStyle
-import androidx.compose.runtime.*
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.awt.ComposeWindow
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.SolidColor
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.window.NoCaretTextField
 import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.rememberWindowState
 import androidx.compose.ui.window.runApplicationTest
 import java.awt.event.KeyEvent
 import java.awt.event.MouseEvent
 import org.junit.Test
-
-@Composable
-fun NoCaretTextField(
-    value: String,
-    onValueChange: (String) -> Unit,
-    modifier: Modifier = Modifier,
-    textStyle: TextStyle = LocalTextStyle.current,
-    visualTransformation: VisualTransformation = VisualTransformation.None,
-) {
-    BasicTextField(
-        value = value,
-        onValueChange = onValueChange,
-        modifier = modifier,
-        textStyle = textStyle,
-        cursorBrush = SolidColor(Color.Unspecified),
-        visualTransformation = visualTransformation,
-    )
-}
 
 class TextFieldUndoTest {
     @Test
